+++ messenger-core/src/main/java/com/example/messenger/message/model/MessageContent.java
@@ -1,11 +1,15 @@
-package com.example.messenger.message;
+package com.example.messenger.message.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
 
-@Data
+@Setter
+@Getter
+@ToString
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class MessageContent {
