+++ messenger-core/src/main/java/com/example/messenger/message/model/MessageContentList.java
@@ -1,11 +1,16 @@
-package com.example.messenger.message;
+package com.example.messenger.message.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
 
 import java.util.List;
 
-@Data
+@Setter
+@Getter
+@ToString
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class MessageContentList {
