+++ messenger-core/src/main/java/com/example/messenger/message/model/MessageDto.java
@@ -1,4 +1,4 @@
-package com.example.messenger.message;
+package com.example.messenger.message.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Value;
@@ -9,12 +9,13 @@ import java.util.UUID;
 /**
  * DTO for {@link MessageEntity}
  */
-@Value
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class MessageDto implements Serializable {
+public record MessageDto(
 
-    UUID id;
-    String content;
-    UUID sessionId;
-    String encryptedAesKey;
-}
+        UUID id,
+
+        String content,
+
+        UUID sessionId,
+
+        String encryptedAesKey) { }
