+++ messenger-core/src/main/java/com/example/messenger/message/model/MessageMapper.java
@@ -1,0 +1,16 @@
+package com.example.messenger.message.model;
+
+import org.mapstruct.Mapper;
+import org.mapstruct.MappingConstants;
+import org.mapstruct.ReportingPolicy;
+import org.mapstruct.factory.Mappers;
+
+@Mapper(unmappedSourcePolicy = ReportingPolicy.IGNORE,
+        unmappedTargetPolicy = ReportingPolicy.IGNORE,
+        componentModel = MappingConstants.ComponentModel.SPRING)
+public interface MessageMapper {
+
+    MessageMapper INSTANCE = Mappers.getMapper(MessageMapper.class);
+
+    MessageDto toDto(MessageEntity messageEntity);
+}
