+++ messenger-core/src/main/java/com/example/messenger/message/service/MessageService.java
@@ -1,4 +1,6 @@
 package com.example.messenger.message;
+import com.example.messenger.message.model.MessageContent;
+import com.example.messenger.message.model.MessageEntity;
 
 import java.util.List;
 import java.util.UUID;
